+++ src/com/app/Account.java
@@ -1,4 +1,4 @@
-package com.charel;
+package com.app;
 
 public class Account {
