+++ src/com/app/Main.java
@@ -1,4 +1,4 @@
-package com.charel;
+package com.app;
 
 public class Main {
