+++ src/com/app/WelcomeScreen.java
@@ -1,4 +1,4 @@
-package com.charel;
+package com.app;
 
 import java.util.Scanner;
